requirements.txt
django>=2.2,<3
django-select2==7.4.2
django-treebeard>=4.3
